* support/ab.c (output_results): Fix gcc warning on LP64 - field   width passed via '*' must be an int.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@819695 13f79535-47bb-0310-9956-ffa450edef68

+++ support/ab.c
@@ -807,7 +807,7 @@ static void output_results(int sig)
     if (send_body)
         printf("Total %s:%*s          %" APR_INT64_T_FMT "\n",
                method_str_pasttense[method],
-               7 - strlen(method_str_pasttense[method]),
+               (int)(7 - strlen(method_str_pasttense[method])),
                " ",
                totalposted);
     printf("HTML transferred:       %" APR_INT64_T_FMT " bytes\n", totalbread);
